Move `Paginator` abstraction to `Helpers`
