+++ Website/ui/src/modules/Client/Ticket.vue
@@ -133,7 +133,7 @@ import Widget from '../../shared/widget'
 import { Ticket, UserTickets } from '@/services/TicketService'
 import { resources } from '@/resources'
 import { EventBus } from '@/shared/eventbus'
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import moment from 'moment'
 import { TicketUserService } from '@/services/TicketUserService'
 import { TicketLabelService } from '@/services/TicketLabelService'
