+++ Website/ui/src/services/AgentService.js
@@ -1,5 +1,5 @@
 import Repository from '../repositories/RepositoryFactory'
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import { EventBus } from '@/shared/eventbus'
 import { ErrorHandler } from '@/Helpers/ErrorHander'
 import { resources } from '@/resources'
