+++ Website/ui/src/services/BookKeepingService.js
@@ -1,5 +1,5 @@
 import Repository from '../repositories/RepositoryFactory'
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import { resources } from '@/resources'
 
 export class BookKeepingService {
