+++ Website/ui/src/services/TicketService.js
@@ -1,6 +1,6 @@
 import Repository from '../repositories/RepositoryFactory'
 import { ErrorHandler } from '@/Helpers/ErrorHander'
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import { resources } from '@/resources'
 import { TicketTrelloService } from './TicketTrelloService'
