+++ Website/ui/src/services/TransactionService.js
@@ -1,5 +1,5 @@
 import Repository from '../repositories/RepositoryFactory'
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import { ErrorHandler } from '@/Helpers/ErrorHander'
 import { resources } from '@/resources'
 import { EventBus } from '@/shared/eventbus'
