+++ Website/ui/src/services/UserTransactionsService.js
@@ -1,6 +1,6 @@
 import Repository from '../repositories/RepositoryFactory'
 import { ErrorHandler } from '@/Helpers/ErrorHander'
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 
 export class UserTransactionsService {
     constructor(personId) {
