+++ Website/ui/src/shared/Paginate.vue
@@ -110,7 +110,7 @@
 </template>
 
 <script>
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import { EventBus } from './eventbus'
 export default {
     name: 'Paginate',
