+++ Website/ui/src/shared/widget.vue
@@ -117,7 +117,7 @@
 </template>
 
 <script>
-import { Paginator } from '@/classes/paginator'
+import { Paginator } from '@/Helpers/Paginator'
 import { EventBus } from './eventbus'
 import Paginate from './Paginate'
